Fix problem when you want to use search with empty string. Also we can now filter by sections while searching

+++ HardwareStore/Controllers/ShoppingCartController.cs
@@ -79,6 +79,7 @@ namespace HardwareStore.Controllers
             _userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
 
             var productToDelete = _context.ShoppingCarts.Where(d => d.IdentityUserId == _userId).SingleOrDefault(d => d.ProductId == id);
+            if (productToDelete == null) return NotFound();
 
             _context.ShoppingCarts.Remove(productToDelete);
